Spectrum: Fork out utility functions into its own class
